requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.3
asttokens==2.4.0
async-generator==1.10
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
behave==1.2.6
certifi==2023.7.22
cffi==1.15.0
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.1.4
cryptography==37.0.2
cssselect==1.2.0
debugpy==1.8.0
decorator==4.4.2
distro==1.8.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
execnet==2.0.2
executing==2.0.0
fasteners==0.19
filelock==3.13.1
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.4
imageio==2.33.0
imageio-ffmpeg==0.4.9
iniconfig==2.0.0
ipykernel==6.25.2
ipython==8.16.1
jedi==0.19.1
jupyter_client==8.4.0
jupyter_core==5.4.0
markdown-it-py==3.0.0
matplotlib-inline==0.1.6
mdurl==0.1.2
moviepy==1.0.3
multidict==6.0.4
nest-asyncio==1.5.8
numpy==1.26.2
openai==1.3.7
openpyxl==3.0.9
outcome==1.3.0.post0
packaging==23.2
parameterized==0.9.0
parse==1.19.1
parse-type==0.6.2
parso==0.8.3
pdbp==1.5.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
pluggy==1.3.0
proglog==0.1.10
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.16.1
pynose==1.4.8
pyOpenSSL==22.0.0
pyotp==2.9.0
PySocks==1.7.1
pytest==7.4.3
pytest-html==2.0.1
pytest-metadata==3.0.0
pytest-ordering==0.6
pytest-rerunfailures==12.0
pytest-xdist==3.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pyzmq==25.1.1
requests==2.31.0
rich==13.6.0
sbvirtualdisplay==1.3.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
tabcompleter==1.3.0
tomli==2.0.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.8.0
urllib3==2.0.7
wcwidth==0.2.8
websockets==12.0
wsproto==1.2.0
yarl==1.9.2
